+++ packages/database/src/core/index.js
@@ -8,3 +8,4 @@ export * from './utilities';
 export { DatabaseModel } from './DatabaseModel';
 export { DatabaseRecord } from './DatabaseRecord';
 export * from './sync';
+export { DatabaseChangeChannel } from './DatabaseChangeChannel';
