+++ packages/database/src/core/modelClasses/Tombstone.js
@@ -1,5 +1,6 @@
 import { SyncDirections } from '@tupaia/constants';
 
+import { DatabaseModel } from '../DatabaseModel';
 import { DatabaseRecord } from '../DatabaseRecord';
 import { RECORDS } from '../records';
