+++ packages/database/src/server/TupaiaChangeChannel.js
@@ -1,0 +1,10 @@
+import PGPubSub from 'pg-pubsub';
+import winston from 'winston';
+
+import { getConnectionConfig } from './getConnectionConfig';
+
+export class TupaiaChangeChannel extends PGPubSub {
+  constructor() {
+    super(getConnectionConfig(), { log: winston.info });
+  }
+}
