+++ packages/database/src/server/TupaiaDatabase.js
@@ -1,13 +1,9 @@
-import winston from 'winston';
 import { types as pgTypes } from 'pg';
 
 import { BaseDatabase } from '../core';
 import { getConnectionConfig } from './getConnectionConfig';
-import { DatabaseChangeChannel } from './DatabaseChangeChannel';
-
-// no math here, just hand-tuned to be as low as possible while
-// keeping all the tests passing
-const HANDLER_DEBOUNCE_DURATION = 250;
+import { DatabaseChangeChannel } from '../core/DatabaseChangeChannel';
+import { TupaiaChangeChannel } from './TupaiaChangeChannel';
 
 export class TupaiaDatabase extends BaseDatabase {
   static IS_CHANGE_HANDLER_SUPPORTED = true;
@@ -33,87 +29,6 @@ export class TupaiaDatabase extends BaseDatabase {
       pgTypes.setTypeParser(pgTypes.builtins.NUMERIC, Number.parseFloat);
       pgTypes.setTypeParser(20, Number.parseInt); // bigInt type to Integer
     }
-  }
-
-  getHandlersForChange(change) {
-    const { handler_key: specificHandlerKey, record_type: recordType } = change;
-    const handlersForCollection = this.getChangeHandlersForCollection(recordType);
-    if (specificHandlerKey) {
-      return handlersForCollection[specificHandlerKey]
-        ? [handlersForCollection[specificHandlerKey]]
-        : [];
-    }
-    return Object.values(handlersForCollection);
-  }
-
-  async notifyChangeHandlers(change) {
-    const unlock = this.handlerLock.createLock(change.record_id);
-    const handlers = this.getHandlersForChange(change);
-    try {
-      for (let i = 0; i < handlers.length; i++) {
-        try {
-          await handlers[i](change);
-        } catch (e) {
-          winston.error(e);
-        }
-      }
-    } finally {
-      unlock();
-    }
-  }
-
-  async waitForAllChangeHandlers() {
-    return this.handlerLock.waitWithDebounce(HANDLER_DEBOUNCE_DURATION);
-  }
-
-  getChangeHandlersForCollection(collectionName) {
-    // Instantiate the array if no change handlers currently exist for the collection
-    if (!this.changeHandlers[collectionName]) {
-      this.changeHandlers[collectionName] = {};
-    }
-    return this.changeHandlers[collectionName];
-  }
-
-  getOrCreateChangeChannel() {
-    if (!this.changeChannel) {
-      this.changeChannel = this.transactingChangeChannel || new DatabaseChangeChannel();
-      this.changeChannel.addDataChangeHandler(this.notifyChangeHandlers);
-      this.changeChannelPromise = this.changeChannel.ping(undefined, 0);
-    }
-    return this.changeChannel;
-  }
-
-  async waitUntilConnected() {
-    await super.waitUntilConnected();
-    if (this.changeChannel) {
-      await this.waitForChangeChannel();
-    }
-  }
-
-  async waitForChangeChannel() {
-    this.getOrCreateChangeChannel();
-    return this.changeChannelPromise;
-  }
-
-  addChangeHandlerForCollection(collectionName, changeHandler, key = this.generateId()) {
-    // if a change handler is being added, this db needs a change channel - make sure it's instantiated
-    this.getOrCreateChangeChannel();
-    this.getChangeHandlersForCollection(collectionName)[key] = changeHandler;
-    return () => {
-      delete this.getChangeHandlersForCollection(collectionName)[key];
-    };
-  }
-
-  async closeConnections() {
-    if (this.changeChannel) {
-      await this.changeChannel.close();
-    }
-    return super.closeConnections();
-  }
-
-  addSchemaChangeHandler(handler) {
-    const changeChannel = this.getOrCreateChangeChannel();
-    return changeChannel.addSchemaChangeHandler(handler);
   }
 
   /**
@@ -147,7 +62,7 @@ export class TupaiaDatabase extends BaseDatabase {
     return new TupaiaDatabase(transaction, this.changeChannel);
   }
 
-  async markRecordsAsChanged(recordType, records) {
-    await this.getOrCreateChangeChannel().publishRecordUpdates(recordType, records);
+  createChangeChannel() {
+    return new DatabaseChangeChannel(new TupaiaChangeChannel());
   }
 }
