+++ packages/database/src/server/index.js
@@ -6,6 +6,5 @@ export * from './testUtilities';
 export { TupaiaDatabase } from './TupaiaDatabase';
 export { getDbMigrator } from './getDbMigrator';
 export { getConnectionConfig } from './getConnectionConfig';
-export { DatabaseChangeChannel } from './DatabaseChangeChannel';
 export { runPostMigration } from './runPostMigration';
 export * from './changeHandlers';
