+++ packages/datatrak-web/src/database/DatatrakDatabase.ts
@@ -1,9 +1,10 @@
 import ClientPgLite from 'knex-pglite';
 import type { Knex } from 'knex';
 
-import { BaseDatabase } from '@tupaia/database';
+import { BaseDatabase, DatabaseChangeChannel } from '@tupaia/database';
 
 import { getConnectionConfig } from './getConnectionConfig';
+import { BrowserChangeChannel } from './BrowserChangeChannel';
 
 /**
  * Ideally this should stay in the database package, but it has to stay here to avoid build problems
@@ -17,5 +18,9 @@ export class DatatrakDatabase extends BaseDatabase {
     return this.connection.transaction(transaction =>
       wrappedFunction(new DatatrakDatabase(transaction)),
     );
+  }
+
+  createChangeChannel() {
+    return new DatabaseChangeChannel(new BrowserChangeChannel());
   }
 }
