+++ packages/datatrak-web/src/database/getConnectionConfig.ts
@@ -1,17 +1,17 @@
-import { PGlite } from '@electric-sql/pglite';
-
 import { getEnvVarOrDefault } from '@tupaia/utils';
+import { getPGliteInstance } from './getPGliteInstance';
 
 export const getConnectionConfig = () => {
-  const connectionString = getEnvVarOrDefault('PG_LITE_CONNECTION_STRING', 'idb://datatrak-db');
-
-  if (process.env.NODE_ENV === 'production') {
-    return {
-      pglite: new PGlite(connectionString),
-    };
-  }
-
   return {
-    connectionString,
+    pglite: getPGliteInstance(),
   };
+  // if (process.env.NODE_ENV === 'production') {
+  //   return {
+  //     pglite: getPGliteInstance(),
+  //   };
+  // }
+
+  // return {
+  //   connectionString: getEnvVarOrDefault('PG_LITE_CONNECTION_STRING', 'idb://datatrak-db'),
+  // };
 };
